Update README.md
In the spirit of assumption,

There has been a high rate of Employees leaving their jobs to relocate to other countries for greener pastures. The Japa Syndrome has hit so bad that employers began to take a survey whenever a staff is leaving. From the many options, I have decided to look into these Five major points as to why they are leaving in the year 2022. The aspects considered in this survey includes: 1. Job was Challenging 2. Open Opportunity 3. Understaffed 4. Skills Maximization 5. Access to Learning/ Development.

Step 1: The visual used is a very beautiful one, the colors are catchy and blends well. The Survey Bars were well represented but a little tedious to understand. A more appealing visual will do justice to the survey.

Step 2: Still on assumption, we can see from the visuals that a very few percentage believe that their Job was Challenging and quite a number believe that the organisation was understaffed. The management will do well to address these Issues as this will lead to more productivity and bring about job satisfaction among members. More Hands should be employed to ease the workload on employees, Skills should be maximised where exposing staffs to more challenging job roles.